Plugins: add opt-out toggle for Live Preview button
This allows plugin committers to disable the Live Preview button for specific plugins. It adds a button to the Danger Zone section on the Advanced view of the plugin page.

See #7251


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@12933 74240141-8908-4e6f-9713-ba540dce6ec7

+++ wordpress.org/public_html/wp-content/themes/pub/wporg-plugins/inc/template-tags.php
@@ -519,6 +519,9 @@ function the_plugin_danger_zone() {
 
 		// Output the self close button.
 		the_plugin_self_close_button();
+
+		// Output the toggle preview button.
+		the_plugin_self_toggle_preview_button();
 	}
 
 }
@@ -553,6 +556,37 @@ function the_plugin_self_close_button() {
 		echo '<form method="POST" action="' . esc_url( $close_link ) . '" onsubmit="return confirm( jQuery(this).prev(\'.notice\').text() );">';
 		// Translators: %s is the plugin name, as defined by the plugin itself.
 		echo '<p><input class="button" type="submit" value="' . esc_attr( sprintf( __( 'I understand, please close %s.', 'wporg-plugins' ), get_the_title() ) ) . '" /></p>';
+		echo '</form>';
+	}
+}
+
+/**
+ * Displays a form for plugin committers to toggle the Live Preview button.
+ */
+function the_plugin_self_toggle_preview_button() {
+	$post            = get_post();
+	$toggle_link     = Template::get_self_toggle_preview_link( $post );
+
+	if ( ! current_user_can( 'plugin_self_close', $post ) ) {
+		return;
+	}
+
+	echo '<h4>' . esc_html__( 'Toggle Live Preview', 'wporg-plugins' ) . '</h4>';
+	$preview_status = get_post_meta( $post->ID, '_no_preview', true ) ? 'disabled' : 'enabled';
+	if ( 'enabled' === $preview_status ) {
+		echo '<p>' . esc_html__( 'The Live Preview link to Playground is currently enabled. Use the toggle button to disable it.', 'wporg-plugins' ) . '</p>';
+	} else {
+		echo '<p>' . esc_html__( 'The Live Preview link to Playground is currently disabled. Use the toggle button to enable it.', 'wporg-plugins' ) . '</p>';
+	}
+
+	echo '<div class="plugin-notice notice notice-warning notice-alt"><p>';
+	_e( '<strong>Note:</strong> This only affects the availability of the Live Preview button on the plugin page. It does not prevent a plugin from running in the Playground.', 'wporg-plugins' );
+	echo '</p></div>';
+
+	if ( $toggle_link ) {
+		echo '<form method="POST" action="' . esc_url( $toggle_link ) . '" onsubmit="return confirm( jQuery(this).prev(\'.notice\').text() );">';
+		// Translators: %s is the plugin name, as defined by the plugin itself.
+		echo '<p><input class="button" type="submit" value="' . esc_attr( sprintf( __( 'Please toggle the Live Preview link for %s', 'wporg-plugins' ), get_the_title() ) ) . '" /></p>';
 		echo '</form>';
 	}
 }
